apps/api/requirements.txt
alembic==1.16.5
faiss-cpu==1.12.0
fastapi==0.115.12
numpy==2.3.3
python-multipart==0.0.20
psycopg[binary]==3.2.9
sqlalchemy==2.0.44
uvicorn[standard]==0.34.2
